APIService code style refactoring
