+++ ionic2/app/models/pokemon-icon.ts
