+++ ionic2/app/models/pokemon.ts
