+++ ionic2/app/services/api.service.ts
@@ -2,12 +2,12 @@ import { Injectable } from '@angular/core';
 import { Http, Response, URLSearchParams } from '@angular/http';
 import { Observable } from 'rxjs';
 
-import { Pokemon } from './models/pokemon';
-import { PokemonIcon } from './models/pokemon-icon';
-import { Sighting } from './models/sighting';
-import {PokemonGender} from "./models/pokemon";
-import {PokemonAttackCategory} from "./models/pokemon";
-import {SightingSource} from "./models/sighting";
+import { Pokemon } from '../models/pokemon';
+import { PokemonIcon } from '../models/pokemon-icon';
+import { Sighting } from '../models/sighting';
+import { PokemonGender } from "../models/pokemon";
+import { PokemonAttackCategory } from "../models/pokemon";
+import { SightingSource } from "../models/sighting";
 
 export { Pokemon, PokemonIcon, Sighting}
